pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "happyhorse-1-api"
version = "0.1.0"
authors = [
  { name="Anil Matcha" },
]
description = "A comprehensive Python wrapper for the HappyHorse 1.0 API by Alibaba."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "python-dotenv",
    "mcp[cli]"
]

[tool.setuptools]
py-modules = ["happyhorse_api", "mcp_server"]
